requirements.txt
pycryptodome==3.21.0
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
pytest-html==4.1.1
PyJWT==2.10.1

